refactor: Renamed `scalebar.ts` to `scaleBar.ts` for consistency
