+++ src/colorizer/canvas/elements/index.ts
@@ -2,12 +2,12 @@ export { defaultFooterStyle, getFooterRenderer } from "./footer";
 export { defaultHeaderStyle, getHeaderRenderer } from "./header";
 export { defaultInsetBoxStyle } from "./insetBox";
 export { defaultLegendStyle } from "./legend";
-export { defaultScaleBarStyle } from "./scalebar";
+export { defaultScaleBarStyle } from "./scaleBar";
 export { defaultTimestampStyle } from "./timestamp";
 
 export type { FooterStyle, FooterParams } from "./footer";
 export type { HeaderStyle, HeaderParams } from "./header";
 export type { InsetBoxStyle } from "./insetBox";
 export type { LegendStyle, LegendParams } from "./legend";
-export type { ScaleBarStyle, ScaleBarParams } from "./scalebar";
+export type { ScaleBarStyle, ScaleBarParams } from "./scaleBar";
 export type { TimestampStyle, TimestampParams } from "./timestamp";
