+++ src/colorizer/canvas/elements/scaleBar.ts
